Correction of anammox and nit-denit processes
In the anammox and nit-denit process (sediment), ammonium is described only as consumed and not produced (the produced ammonium is subtracted from the consumed when unifying the 2-stage process). However, from a programming point of view it is more convenient to consider both the anammox and nit-denit as mineralization processes and so leave the ammonium production. This fixed a mistake in ammonium production and oxygen consumption. In addition, a more significant mistake was detected – a wrong sign in the nit-denit process! this significantly affects sediment interactions! 
Ammonium consumption in nit-denit process is now:
oxb_switch * recs * ( 1.0_rk + 4.3_rk * self%fds)
and oxygen consumption (moved to mineralization rate, line 407) is:
recs * oxb_switch * (1.0_rk + 0.6_rk * self%fds)