fixes and Anammox process in sediment when water is oxic
I added anammox also to the sediment when water is with oxygen. this is similar logic to nitrification-denitrification process, but nitrification-anammox process. The ratio between denitrification and anammox is controlled by the same variable - den_frac_denabmx_sed.
yet again, I found significant errors in my previous code - ammonium was consumed also for the oxic part of the sediment. It is now fixed. Documentation is being prepared.
I also deleted (turned to comment) the aggregation of sediment parts of fluf into total mass because it is mmole/m2 and not mmole/m3.